+++ src/main/java/net/openhft/chronicle/bytes/StreamingDataInput.java
@@ -282,7 +282,7 @@ public interface StreamingDataInput<S extends StreamingDataInput<S>> extends Str
         if (len0 == -1)
             return false;
         int len = Maths.toUInt31(len0);
-        b.write((BytesStore) this, readPosition(), (long) len);
+        b.write((BytesStore) this, readPosition(), len);
         readSkip(len);
         return true;
     }
